requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
deprecation==2.1.0
idna==3.4
importlib-metadata==6.0.0
mashumaro==3.5
packaging==23.0
py==1.11.0
pyparsing==3.0.9
pysnyk==0.9.5
requests==2.28.2
retry==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.13.0
